+++ src/HiGlassMetaConsumer.js
@@ -3,8 +3,6 @@ import isEqual from 'lodash/isEqual';
 import clamp from 'lodash/clamp';
 
 import { HiGlassComponent } from 'higlass';
-import higlassRegister from 'higlass-register';
-import StackedBarTrack from 'higlass-multivec/es/StackedBarTrack.js';
 
 import { InfoContext, ACTION } from './utils/contexts.js';
 import { selectRows, highlightRowsFromSearch } from './utils/select-rows.js';
@@ -40,12 +38,6 @@ import { wrapSvg } from './utils/wrap-svg.js';
 import './HiGlassMetaConsumer.scss';
 import cloneDeep from 'lodash/cloneDeep';
 import { removeItemFromArray, modifyItemInArray, insertItemToArray } from './utils/array.js';
-
-higlassRegister({
-    name: 'StackedBarTrack',
-    track: StackedBarTrack,
-    config: StackedBarTrack.config,
-});
 
 const hgOptionsBase = {
     sizeMode: 'bounded', // Stretch the height of HiGlass to its container <div/>
@@ -85,7 +77,8 @@ const HiGlassMetaConsumer = forwardRef((props, ref) => {
         ref.current = {
             api: {
                 onOptions: (newOptions) => setOptions(processWrapperOptions(newOptions)),
-                onRemoveAllFilters: () => removeAllFilters()
+                onRemoveAllFilters: () => removeAllFilters(),
+                onRemoveAllSort: () => removeAllSort()
             }
         }
     }, [ref, multivecTrackIds]);
@@ -284,10 +277,33 @@ const HiGlassMetaConsumer = forwardRef((props, ref) => {
         setOptions(newOptions);
     }, [multivecTrackIds]);
 
+    const removeAllSort = useCallback(() => {
+        let newOptions = options;
+        multivecTrackIds.forEach(({ viewId, trackId }) => {
+            newOptions = updateWrapperOptions(newOptions, [], "rowSort", viewId, trackId, { isReplace: true });
+        });
+        setOptions(newOptions);
+    }, [multivecTrackIds]);
+
     // Callback function for sorting.
-    const onSortRows = useCallback((viewId, trackId, field, type, order) => {
-        const newRowSort = [ { field, type, order } ];
-        const newOptions = updateWrapperOptions(options, newRowSort, "rowSort", viewId, trackId, { isReplace: true });
+    const onSortRows = useCallback((viewId, trackId, field, type, order, isTrackIndependent) => {
+        let newOptions;
+        if(isTrackIndependent) {
+            // We need to modify the track-level `sort` option
+            const { rowInfoAttributes } = getTrackWrapperOptions(options, viewId, trackId);
+            const newRowInfoAttributes = rowInfoAttributes.map(fieldInfo => {
+                if(fieldInfo.field === field && fieldInfo.type === type) {
+                    return { ...fieldInfo, sort: order }
+                } else {
+                    return fieldInfo;
+                }
+            });
+            newOptions = updateWrapperOptions(options, newRowInfoAttributes, "rowInfoAttributes", viewId, trackId, { isReplace: true });
+        }
+        else {
+            const newRowSort = [ { field, type, order } ];
+            newOptions = updateWrapperOptions(options, newRowSort, "rowSort", viewId, trackId, { isReplace: true });
+        }
         setOptions(newOptions);
     }, [options]);
 
@@ -396,7 +412,7 @@ const HiGlassMetaConsumer = forwardRef((props, ref) => {
         setSelectedRowsToViewConfig(viewId, trackId, newSelectedRows);
         setOptions(newWrapperOptions);
     }, [options]);
-
+    
     // Callback function for adding a track.
     const onAddTrack = useCallback((viewId, trackId, field, type, notOneOf, position) => {
         if(viewId === DEFAULT_OPTIONS_KEY || trackId === DEFAULT_OPTIONS_KEY) {
@@ -524,8 +540,8 @@ const HiGlassMetaConsumer = forwardRef((props, ref) => {
                     onAddTrack={(field, type, notOneOf, position) => {
                         onAddTrack(viewId, trackId, field, type, notOneOf, position);
                     }}
-                    onSortRows={(field, type, order) => {
-                        onSortRows(viewId, trackId, field, type, order);
+                    onSortRows={(field, type, order, isTrackIndependent) => {
+                        onSortRows(viewId, trackId, field, type, order, isTrackIndependent);
                     }}
                     onHighlightRows={(field, type, condition) => {
                         onHighlightRows(viewId, trackId, field, type, condition);