+++ src/TrackRowInfoVis.js
@@ -5,12 +5,15 @@ import TrackRowInfoVisNominalBar from './TrackRowInfoVisNominalBar.js';
 import TrackRowInfoVisQuantitativeBar from './TrackRowInfoVisQuantitativeBar.js';
 import TrackRowInfoVisLink from './TrackRowInfoVisLink.js';
 import TrackRowInfoVisDendrogram from './TrackRowInfoVisDendrogram.js';
+import TrackRowInfoVisBand from './TrackRowInfoVisBand.js';
+import { HIGLASSMETA_DEFAULT } from './utils/visualization-properties.js';
 
 const fieldTypeToVisComponent = {
     "nominal": TrackRowInfoVisNominalBar,
     "quantitative": TrackRowInfoVisQuantitativeBar,
     "url": TrackRowInfoVisLink,
-    "tree": TrackRowInfoVisDendrogram
+    "tree": TrackRowInfoVisDendrogram,
+    "band": TrackRowInfoVisBand
 };
 
 /**
@@ -21,6 +24,10 @@ const fieldTypeToVisComponent = {
  * @prop {number} height The height of this view.
  * @prop {object[]} rowInfo The array of JSON Object containing row information.
  * @prop {object[]} transformedRowInfo The `rowInfo` array after aggregating, filtering, and sorting rows.
+ * @prop {array} leftSelectedRows The array of selected indices on the left track.
+ * @prop {array} rightSelectedRows The array of selected indices on the right track.
+ * @prop {array} selectedRows The array of selected indices. 
+ * @prop {array} highlitRows The array of highlit indices.
  * @prop {object} fieldInfo The name and type of data field.
  * @prop {boolean} isLeft Is this view on the left side of the track?
  * @prop {object} rowSort The options for sorting rows.
@@ -40,6 +47,10 @@ export default function TrackRowInfoVis(props) {
         isLeft,
         rowInfo,
         transformedRowInfo,
+        leftSelectedRows,
+        rightSelectedRows,
+        selectedRows,
+        highlitRows,
         rowSort,
         rowFilter,
         rowHighlight,
@@ -51,7 +62,6 @@ export default function TrackRowInfoVis(props) {
         onWidthChanged
     } = props;
 
-    const minWidth = 40;
     const resizerWidth = 4
     const resizerHeight = 10
     const resizerMargin = 2;
@@ -76,8 +86,8 @@ export default function TrackRowInfoVis(props) {
         const event = d3.event;
         const diff = event.sourceEvent.clientX - dragX.current;
         let newWidth = isLeft ? width - diff : width + diff;
-        if(newWidth < minWidth) {
-            newWidth = minWidth;
+        if(newWidth < HIGLASSMETA_DEFAULT.TRACK.MIN_WIDTH) {
+            newWidth = HIGLASSMETA_DEFAULT.TRACK.MIN_WIDTH;
         }
         // Emit the new width value to the parent component.
         onWidthChanged(newWidth);
@@ -158,6 +168,10 @@ export default function TrackRowInfoVis(props) {
                     fieldInfo,
                     transformedRowInfo,
                     rowInfo,
+                    leftSelectedRows,
+                    rightSelectedRows,
+                    selectedRows,
+                    highlitRows,
                     titleSuffix,
                     sortInfo,
                     filterInfo,