+++ src/TrackRowInfoVisLink.js
@@ -9,6 +9,7 @@ import { drawVisTitle } from "./utils/vis.js";
 import { TooltipContent, destroyTooltip } from "./Tooltip.js";
 import TrackRowInfoControl from './TrackRowInfoControl.js';
 import { getAggregatedValue } from "./utils/aggregate.js";
+import { drawRowHighlightRect } from "./utils/linking.js";
 
 const margin = 5;
 
@@ -20,6 +21,8 @@ const margin = 5;
  * @prop {number} height The height of this view.
  * @prop {object[]} rowInfo The array of JSON Object containing row information.
  * @prop {object[]} transformedRowInfo The `rowInfo` array after aggregating, filtering, and sorting rows.
+ * @prop {array} selectedRows The array of selected indices. 
+ * @prop {array} highlitRows The array of highlit indices.
  * @prop {object} fieldInfo The name and type of data field.
  * @prop {boolean} isLeft Is this view on the left side of the track?
  * @prop {string} titleSuffix The suffix of a title, information about sorting and filtering status.
@@ -39,6 +42,8 @@ export default function TrackRowInfoVisLink(props) {
         isShowControlButtons,
         rowInfo,
         transformedRowInfo,
+        selectedRows,
+        highlitRows,
         titleSuffix,
         sortInfo,
         filterInfo,
@@ -77,12 +82,6 @@ export default function TrackRowInfoVisLink(props) {
 
         const shouldRenderText = (rowHeight >= fontSize);
 
-        if(hoverIndex !== null) {
-            // There is currently a hovered element, so render a background rect.
-            const bgRect = two.makeRect(0, yScale(hoverIndex), width, rowHeight);
-            bgRect.fill = "#EBEBEB";
-        }
-
         if(shouldRenderText) {
             // There is enough height to render the text elements.
             transformedRowInfo.forEach((info, i) => {
@@ -109,8 +108,11 @@ export default function TrackRowInfoVisLink(props) {
             });
         }
         
+        drawRowHighlightRect(two, selectedRows, highlitRows, width, height);
 
-        drawVisTitle(field, { two, isLeft, width, height, titleSuffix });
+        if(!isShowControlButtons) {
+            drawVisTitle(field, { two, isLeft, width, height, titleSuffix });
+        }
         
         two.update();
         return two.teardown;
@@ -131,9 +133,11 @@ export default function TrackRowInfoVisLink(props) {
             if(y !== undefined) {
                 fieldVal = aggValue(transformedRowInfo[y], field);
                 setHoverIndex(y);
+                onHighlightRows(field, "nominal", fieldVal);
             } else {
                 setHoverIndex(null);
                 destroyTooltip();
+                onHighlightRows("");
                 return;
             }
 
@@ -162,13 +166,16 @@ export default function TrackRowInfoVisLink(props) {
 
         // Handle mouse leave.
         d3.select(canvas).on("mouseout", destroyTooltip);
-        d3.select(div).on("mouseleave", () => setHoverIndex(null));
+        d3.select(div).on("mouseleave", () => {
+            setHoverIndex(null);
+            onHighlightRows("");
+        });
 
         return () => {
             teardown();
             d3.select(div).on("mouseleave", null);
         };
-    }, [top, left, width, height, transformedRowInfo, hoverIndex]);
+    }, [top, left, width, height, transformedRowInfo, hoverIndex, isShowControlButtons]);
 
     return (
         <div