+++ src/utils/visualization-properties.js
@@ -1,0 +1,6 @@
+/* Default Visualization Properties */
+export const HIGLASSMETA_DEFAULT = Object.freeze({
+    TRACK: {
+        MIN_WIDTH: 24
+    }
+})